requirements.txt
PyYAML
requests
selenium
